common function when generate packages files
